agents-backend/agents/requirements.txt
requests
psycopg2-binary
pandas
pyyaml
defog>=0.56.4
snowflake-connector-python
pandas
sqlalchemy
scipy
tiktoken
fastapi
scikit-survival
numpy
google-cloud-storage
databricks-sql-connector
pika
hypercorn
colorama
matplotlib
seaborn
redis